backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
fastapi-mail
google-auth
google-auth-oauthlib
requests
playwright
beautifulsoup4
apscheduler
structlog